+++ harbor_cli/__init__.py
@@ -2,6 +2,9 @@
 #
 # SPDX-License-Identifier: MIT
 from __future__ import annotations
+from ._patches import patch_all
+
+patch_all()
 
 from . import logs  # type: ignore # noreorder # configure logger first as side-effect
 from . import *